(feat) renders bare-bones Spinner when photo uploads
